requirements.txt
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiomysql==0.2.0
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.3.0
azure-cognitiveservices-speech==1.42.0
bottle==0.13.2
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
defusedxml==0.7.1
exceptiongroup==1.2.2
fastapi==0.115.6
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
iso8601==2.1.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
mysql-connector-python==9.1.0
networkx==3.4.2
openai==1.57.4
packaging==24.2
pydantic==2.10.4
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
SQLAlchemy==2.0.37
starlette==0.41.3
threadpoolctl==3.5.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
yarl==1.18.3
faiss-cpu==1.9.0.post1
